Refactor chat panel composer ownership
